mcp/requirements.txt
# MCP服务器依赖
Flask>=3.0.0
requests>=2.31.0
pymysql>=1.1.0  # MySQL客户端库（可选：也可使用 mysql-connector-python）

# LangGraph集成相关（可选）
langchain>=1.0.0
langchain-openai>=1.0.0
langchain-core>=1.0.0
langgraph>=1.0.0
python-dotenv>=1.0.0
pydantic>=1.10.0  # Pydantic 模型支持（LangChain 可能需要）

